ref: Enabling drag-n-drop exe
When a user moves the exe into their parse dir and runs it for the first
time, this is known in this project as `drag-n-drop exe`.
It should create the config, default html and styles and generate the
site.
These changes move the default html and css into memory because they won't
be available to the process in a drag-n-drop context.
Also automatically ignore the outDir or it will loop forever.