fix: Remove pagination
Now that `publish` is a metadata and that metadata isn't processed
until each file is processed, non-published files are showing up in
the `next` button.  There's no good solution to this currently and
also most digital gardens don't read well like a book so for now
I'm removing it entirely.